convert env state to np.float32 to fix state not contained in observation space error

+++ robo_gym/envs/ur/ur_avoidance_iros.py
@@ -192,7 +192,7 @@ class AvoidanceIros2021UR(URBaseAvoidanceEnv):
 
         state = np.concatenate((state, trajectory_joint_position, [target_point_flag]))
 
-        return state
+        return state.astype(np.float32)
 
     def _get_observation_space(self) -> gym.spaces.Box:
         """Get environment observation space."""
